requirements.txt
asgiref==3.8.1
crispy-bootstrap5==0.7
Django==4.2.11
django-crispy-forms==2.0
django-crontab==0.7.1
django-filter==24.2
pillow==10.3.0
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
